Transfer getGroupsRenderedData to utils
