+++ js/renovation/ui/scheduler/workspaces/base/group_panel/__tests__/group_panel.test.tsx
@@ -6,6 +6,11 @@ import {
 import { GroupPanelVerticalLayout } from '../vertical/layout';
 import { VERTICAL_GROUP_ORIENTATION, HORIZONTAL_GROUP_ORIENTATION } from '../../../../consts';
 import { GroupPanelHorizontalLayout } from '../horizontal/layout';
+import { getGroupsRenderData } from '../utils';
+
+jest.mock('../utils', () => ({
+  getGroupsRenderData: jest.fn(() => 'data'),
+}));
 
 describe('GroupPanel Vertical Layout', () => {
   describe('Render', () => {
@@ -75,190 +80,21 @@ describe('GroupPanel Vertical Layout', () => {
         });
       });
 
-      describe('groupsRenderData', () => {
-        const groupsBase = [{
-          name: 'group 1',
-          items: [{
-            text: 'item 1', id: 1, color: 'color 1',
-          }, {
-            text: 'item 2', id: 2, color: 'color 2',
-          }],
-          data: [{
-            text: 'item 1', id: 1, color: 'color 1',
-          }, {
-            text: 'item 2', id: 2, color: 'color 2',
-          }],
-        }, {
-          name: 'group 2',
-          items: [{
-            text: 'item 3', id: 1, color: 'color 3',
-          }, {
-            text: 'item 4', id: 2, color: 'color 4',
-          }],
-          data: [{
-            text: 'item 3', id: 1, color: 'color 3',
-          }, {
-            text: 'item 4', id: 2, color: 'color 4',
-          }],
-        }];
-
-        it('should transform grouping data into group items', () => {
-          const groupPanel = new GroupPanel({
-            groups: groupsBase,
-          });
-
-          expect(groupPanel.groupsRenderData)
-            .toEqual([[{
-              ...groupsBase[0].items[0],
-              data: groupsBase[0].data[0],
-              resourceName: groupsBase[0].name,
-              key: '0_group 1_1',
-            }, {
-              ...groupsBase[0].items[1],
-              data: groupsBase[0].data[1],
-              resourceName: groupsBase[0].name,
-              key: '0_group 1_2',
-            }], [{
-              ...groupsBase[1].items[0],
-              data: groupsBase[1].data[0],
-              resourceName: groupsBase[1].name,
-              key: '0_group 2_1',
-            }, {
-              ...groupsBase[1].items[1],
-              data: groupsBase[1].data[1],
-              resourceName: groupsBase[1].name,
-              key: '0_group 2_2',
-            }, {
-              ...groupsBase[1].items[0],
-              data: groupsBase[1].data[0],
-              resourceName: groupsBase[1].name,
-              key: '1_group 2_1',
-            }, {
-              ...groupsBase[1].items[1],
-              data: groupsBase[1].data[1],
-              resourceName: groupsBase[1].name,
-              key: '1_group 2_2',
-            }]]);
-        });
-
-        it('should work when data parameter is undefined', () => {
-          const groups = [{
-            name: 'group 1',
-            items: [{
-              text: 'item 1', id: 1, color: 'color 1',
-            }, {
-              text: 'item 2', id: 2, color: 'color 2',
-            }],
-          }] as any;
+      describe('groupsRenderedData', () => {
+        it('should call getGroupsRenderedData with correct parameters', () => {
+          const groups = [];
           const groupPanel = new GroupPanel({
             groups,
-          });
-
-          expect(groupPanel.groupsRenderData)
-            .toEqual([[{
-              ...groups[0].items[0],
-              resourceName: groups[0].name,
-              key: '0_group 1_1',
-            }, {
-              ...groups[0].items[1],
-              resourceName: groups[0].name,
-              key: '0_group 1_2',
-            }]]);
-        });
-
-        it('should transform grouping data into group items corectly when appointments are groupped by date', () => {
-          const groupPanel = new GroupPanel({
-            groups: groupsBase,
             groupByDate: true,
-            columnCountPerGroup: 2,
+            columnCountPerGroup: 72,
           });
 
           expect(groupPanel.groupsRenderData)
-            .toEqual([[{
-              ...groupsBase[0].items[0],
-              data: groupsBase[0].data[0],
-              resourceName: groupsBase[0].name,
-              key: '0_group 1_1_group_by_date_0',
-              isFirstGroupCell: true,
-              isLastGroupCell: false,
-            }, {
-              ...groupsBase[0].items[1],
-              data: groupsBase[0].data[1],
-              resourceName: groupsBase[0].name,
-              key: '0_group 1_2_group_by_date_0',
-              isFirstGroupCell: false,
-              isLastGroupCell: true,
-            }, {
-              ...groupsBase[0].items[0],
-              data: groupsBase[0].data[0],
-              resourceName: groupsBase[0].name,
-              key: '0_group 1_1_group_by_date_1',
-              isFirstGroupCell: true,
-              isLastGroupCell: false,
-            }, {
-              ...groupsBase[0].items[1],
-              data: groupsBase[0].data[1],
-              resourceName: groupsBase[0].name,
-              key: '0_group 1_2_group_by_date_1',
-              isFirstGroupCell: false,
-              isLastGroupCell: true,
-            }], [{
-              ...groupsBase[1].items[0],
-              data: groupsBase[1].data[0],
-              resourceName: groupsBase[1].name,
-              key: '0_group 2_1_group_by_date_0',
-              isFirstGroupCell: true,
-              isLastGroupCell: false,
-            }, {
-              ...groupsBase[1].items[1],
-              data: groupsBase[1].data[1],
-              resourceName: groupsBase[1].name,
-              key: '0_group 2_2_group_by_date_0',
-              isFirstGroupCell: false,
-              isLastGroupCell: false,
-            }, {
-              ...groupsBase[1].items[0],
-              data: groupsBase[1].data[0],
-              resourceName: groupsBase[1].name,
-              key: '1_group 2_1_group_by_date_0',
-              isFirstGroupCell: false,
-              isLastGroupCell: false,
-            }, {
-              ...groupsBase[1].items[1],
-              data: groupsBase[1].data[1],
-              resourceName: groupsBase[1].name,
-              key: '1_group 2_2_group_by_date_0',
-              isFirstGroupCell: false,
-              isLastGroupCell: true,
-            }, {
-              ...groupsBase[1].items[0],
-              data: groupsBase[1].data[0],
-              resourceName: groupsBase[1].name,
-              key: '0_group 2_1_group_by_date_1',
-              isFirstGroupCell: true,
-              isLastGroupCell: false,
-            }, {
-              ...groupsBase[1].items[1],
-              data: groupsBase[1].data[1],
-              resourceName: groupsBase[1].name,
-              key: '0_group 2_2_group_by_date_1',
-              isFirstGroupCell: false,
-              isLastGroupCell: false,
-            }, {
-              ...groupsBase[1].items[0],
-              data: groupsBase[1].data[0],
-              resourceName: groupsBase[1].name,
-              key: '1_group 2_1_group_by_date_1',
-              isFirstGroupCell: false,
-              isLastGroupCell: false,
-            }, {
-              ...groupsBase[1].items[1],
-              data: groupsBase[1].data[1],
-              resourceName: groupsBase[1].name,
-              key: '1_group 2_2_group_by_date_1',
-              isFirstGroupCell: false,
-              isLastGroupCell: true,
-            }]]);
+            .toBe('data');
+          expect(getGroupsRenderData)
+            .toBeCalledWith(
+              groups, 72, true,
+            );
         });
       });
     });
