+++ js/renovation/ui/scheduler/workspaces/base/group_panel/group_panel.tsx
@@ -8,56 +8,8 @@ import { GroupPanelProps } from './group_panel_props';
 import { GroupPanelVerticalLayout } from './vertical/layout';
 import { GroupPanelHorizontalLayout } from './horizontal/layout';
 import { GroupPanelLayoutProps } from './group_panel_layout_props';
-import {
-  GroupRenderItem,
-  Group,
-  GroupItem,
-} from '../../types.d';
-
-const getGroupsRenderData = (
-  groups: Group[],
-  columnCountPerGroup: number,
-  groupByDate: boolean,
-): GroupRenderItem[][] => {
-  let repeatCount = 1;
-  let groupRenderItems = groups.map((group: Group) => {
-    const result = [] as GroupRenderItem[];
-    const { name: resourceName, items, data } = group;
-
-    for (let i = 0; i < repeatCount; i += 1) {
-      result.push(...items.map(({ id, text, color }: GroupItem, index: number) => ({
-        id,
-        text,
-        color,
-        key: `${i}_${resourceName}_${id}`,
-        resourceName,
-        data: data?.[index],
-      })));
-    }
-
-    repeatCount *= items.length;
-    return result;
-  });
-
-  if (groupByDate) {
-    groupRenderItems = [...(new Array(columnCountPerGroup))]
-      .reduce((currentGroupItems, _, index) => groupRenderItems.map((groupsRow, rowIndex) => {
-        const currentRow = currentGroupItems[rowIndex] || [];
-
-        return [
-          ...currentRow,
-          ...groupsRow.map((item, columnIndex) => ({
-            ...item,
-            key: `${item.key}_group_by_date_${index}`,
-            isFirstGroupCell: columnIndex === 0,
-            isLastGroupCell: columnIndex === groupsRow.length - 1,
-          })),
-        ];
-      }), []);
-  }
-
-  return groupRenderItems;
-};
+import { GroupRenderItem } from '../../types.d';
+import { getGroupsRenderData } from './utils';
 
 export const viewFunction = ({
   layout: Layout,
